refactor: divide pipeline._io into a package
