+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_job/pipeline/_io/__init__.py
@@ -1,0 +1,21 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+"""Classes in this package converts input & output set by user to pipeline job input & output."""
+
+from .base import PipelineOutput, PipelineInput, NodeOutput, NodeInput, InputOutputBase
+from .attr_dict import OutputsAttrDict, _GroupAttrDict
+from .mixin import PipelineNodeIOMixin, AutoMLNodeIOMixin, PipelineIOMixin
+
+__all__ = [
+    "PipelineOutput",
+    "PipelineInput",
+    "NodeOutput",
+    "NodeInput",
+    "InputOutputBase",
+    "OutputsAttrDict",
+    "_GroupAttrDict",
+    "PipelineNodeIOMixin",
+    "AutoMLNodeIOMixin",
+    "PipelineIOMixin",
+]
